DN-4: Refactored comments into modules and created API status component.
